Stage files for blast append

+++ src/itaxotools/blastax/tasks/common/process.py
@@ -36,19 +36,21 @@ def stage_paths(
     input_paths: list[Path],
     output_paths: list[Path],
     db_paths: list[Path] = None,
+    dry: bool = False,
 ) -> dict[Path, Path]:
     staged_paths: dict[Path, Path] = IdentityDict()
     input_dir = work_dir / "input"
     output_dir = work_dir / "output"
-    input_dir.mkdir()
-    output_dir.mkdir()
+    input_dir.mkdir(exist_ok=True)
+    output_dir.mkdir(exist_ok=True)
     for path in input_paths:
         if not _is_str_safe(str(path)):
             staged_path = input_dir / make_str_blast_safe(path.name)
             while staged_path in staged_paths.values():
                 staged_path = staged_path.with_stem(staged_path.stem + choice(ascii_letters))
             staged_paths[path] = staged_path
-            shutil.copy(path, staged_paths[path])
+            if not dry:
+                shutil.copy(path, staged_paths[path])
     for path in output_paths:
         if not _is_str_safe(str(path)):
             if path.exists() and path.is_dir():
@@ -67,17 +69,20 @@ def stage_paths(
                 staged_paths[db_path] = staged_path
                 for path in _get_database_paths(db_path):
                     staged_paths[path] = Path(input_dir / staged_path.name).with_suffix(path.suffix)
-                    shutil.copy(path, staged_paths[path])
+                    if not dry:
+                        shutil.copy(path, staged_paths[path])
 
     return staged_paths
 
 
-def unstage_paths(work_dir: Path, staged_paths: dict[Path, Path], output_path: Path):
+def unstage_paths(work_dir: Path, staged_paths: dict[Path, Path], output_path: Path = None, clear: bool = True):
     input_dir = work_dir / "input"
     output_dir = work_dir / "output"
-    if staged_paths[output_path].is_file():
-        shutil.copy(staged_paths[output_path], output_path)
-    else:
-        shutil.copytree(output_dir, output_path, dirs_exist_ok=True)
-    shutil.rmtree(output_dir)
-    shutil.rmtree(input_dir)
+    if output_path is not None:
+        if staged_paths[output_path].is_file():
+            shutil.copy(staged_paths[output_path], output_path)
+        else:
+            shutil.copytree(output_dir, output_path, dirs_exist_ok=True)
+    if clear:
+        shutil.rmtree(output_dir)
+        shutil.rmtree(input_dir)
